the core is transferred to src and the type association is modified using the .d.ts files
