+++ src/core.js
@@ -1,4 +1,3 @@
 export * from "./element/custom-element";
-export * from "./element/hooks";
-export * from "./hooks";
+export * from "./hooks/hooks";
 export * from "./render";
