--- src/core/element/hooks.js
@@ -1,8 +1,0 @@
-export * from "./hooks/use-prop";
-export * from "./hooks/use-event";
-/**
- * @deprecated
- * export * from "./hooks/use-style-sheet";
- * su utilidad no es argumento para acoplar dentro del núcleo, se nueve a un módulo externo
- * export * from "./hooks/use-delegate-focus";
- */
