+++ src/html.js
@@ -1,5 +1,5 @@
 import htm from "htm";
-import { h } from "./core/core";
+import { h } from "./core";
 
 export const html = htm.bind(h);
 export default html;
