+++ src/internal.d.ts
@@ -1,0 +1,3 @@
+export interface flatParamMap<T> extends Array<any> {
+    _?: Map<any, any>;
+}
