+++ src/utils.js
@@ -1,12 +1,13 @@
 /**
  * compare 2 array
- * @param {array} before
- * @param {array} after
- * @example
+ * ```js
  * isEqualArray([1,2,3,4],[1,2,3,4]) // true
  * isEqualArray([1,2,3,4],[1,2,3])   // false
  * isEqualArray([5,1,2,3],[1,2,3,5]) // false
  * isEqualArray([],[]) // true
+ * ```
+ * @param {Array<any>} before
+ * @param {Array<any>} after
  * @returns {boolean}
  */
 export function isEqualArray(before, after) {
@@ -17,7 +18,14 @@ export function isEqualArray(before, after) {
     }
     return true;
 }
-
+/**
+ * Determine if the value is considered a function
+ * @param {any} value
+ */
 export const isFunction = (value) => typeof value == "function";
 
+/**
+ * Determines if the value is considered an object
+ * @param {any} value
+ */
 export const isObject = (value) => typeof value == "object";
